% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_finalise_plot.R
\name{create_footer}
\alias{create_footer}
\title{Helper function for \code{finalise_plot()}, create footer for your plot}
\usage{
create_footer(source_name)
}
\arguments{
\item{source_name}{The text you want to come after the text 'Source:'
in the bottom left hand side of your side}
}
\description{
Helper function for \code{finalise_plot()}, create footer for your plot
}
