% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_finalise_plot.R
\name{finalise_ggplot}
\alias{finalise_ggplot}
\title{Arrange alignment and save SCC ggplot chart}
\usage{
finalise_ggplot(
  plot_name,
  source_name,
  save_filepath = file.path(Sys.getenv("TMPDIR"), "tmp-nc.png"),
  width_pixels = 640,
  height_pixels = 450
)
}
\arguments{
\item{plot_name}{The variable name of the plot you have created that you
want to format and save}

\item{source_name}{The text you want to come after the text 'Source:'
in the bottom left hand side of your side}

\item{save_filepath}{Exact filepath that you want the plot to be saved to}

\item{width_pixels}{Width in pixels that you want to save your chart to -
defaults to 640}

\item{height_pixels}{Height in pixels that you want to save your chart to -
defaults to 450
which needs to be a PNG file}
}
\description{
Running this function will save your plot with the correct guidelines
for publication for a SCC graphic. It will left align your title,
subtitle and source and save it to your specified location.
}
\examples{
\dontrun{
finalise_ggplot(
  plot_name = myplot,
  source = "The source for my data",
  save_filepath = "filename_that_my_plot_should_be_saved_to-nc.png",
  width_pixels = 640,
  height_pixels = 450,
)
}

}
