% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_finalise_plot.R
\name{left_align_text}
\alias{left_align_text}
\title{Helper function for \code{finalise_ggplot()}, left align text}
\usage{
left_align_text(plot_name, pieces)
}
\arguments{
\item{plot_name}{The variable name of the plot you have created that you
want to format and save}

\item{pieces}{The pieces of your plot, ie Title, subtitle and caption}
}
\description{
Helper function for \code{finalise_ggplot()}, left align text
}
