% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_finalise_plot.R
\name{save_plot}
\alias{save_plot}
\title{Helper function for \code{finalise_ggplot()}, saving your plot}
\usage{
save_plot(plot_grid, width_pixels, height_pixels, save_filepath)
}
\arguments{
\item{plot_grid}{The variable name of the plot you have created that you
want to format and save}

\item{width_pixels}{Width in pixels that you want to save your chart to}

\item{height_pixels}{Height in pixels that you want to save your chart to}

\item{save_filepath}{Exact filepath that you want the plot to be saved to}
}
\description{
Helper function for \code{finalise_ggplot()}, saving your plot
}
