% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_barchart.r
\name{scc_barchart}
\alias{scc_barchart}
\title{Define barchart in scc colours}
\usage{
scc_barchart(
  df_to_plot,
  x,
  y,
  title = NULL,
  subtitle = NULL,
  format = "report"
)
}
\arguments{
\item{df_to_plot}{The dataframe provided}

\item{x}{The x-axis variable}

\item{y}{The y-axis variable}

\item{title}{The title that should go above the plot}

\item{subtitle}{The subtitle that should go below the title}

\item{format}{can be one of "report", "presentation" or "tweet". Will adjust
font sizes approriately for that format.}
}
\description{
Define barchart in scc colours
}
