% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_barchart.r
\name{scc_barchart_grouped}
\alias{scc_barchart_grouped}
\title{Define a grouped barchart in scc colours}
\usage{
scc_barchart_grouped(data, x, y, group, title)
}
\arguments{
\item{data}{The dataframe provided}

\item{x}{The x-axis variable}

\item{y}{The y-axis variable}

\item{group}{the variable by which the chart should be grouped}

\item{title}{The title that should go above the plot}
}
\description{
Define a grouped barchart in scc colours
}
