% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_barchart.r
\name{scc_barchart_pupils}
\alias{scc_barchart_pupils}
\title{Define barchart in scc colours, specifically for pupil/population numbers}
\usage{
scc_barchart_pupils(data, x, y, title)
}
\arguments{
\item{data}{The dataframe provided}

\item{x}{The x-axis variable}

\item{y}{The y-axis variable}

\item{title}{The title that should go above the plot}
}
\description{
Define barchart in scc colours, specifically for pupil/population numbers
}
