% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_piechart.R
\name{scc_piechart}
\alias{scc_piechart}
\title{Define piechart in scc colours}
\usage{
scc_piechart(df, value, group, title = NULL, subtitle = NULL)
}
\arguments{
\item{df}{The dataframe provided}

\item{value}{The value of each group, no need to convert into a percentage}

\item{group}{The associated group for each value}

\item{title}{The title that should go above the plot}

\item{subtitle}{The subtitle that should go below the title}
}
\description{
Define piechart in scc colours
}
