% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_style.r
\name{scc_style}
\alias{scc_style}
\title{Defince scc style for ggplot2}
\usage{
scc_style(format = "report")
}
\arguments{
\item{format}{can be one of "report", "presentation" or "tweet". Will adjust
font sizes approriately for that format.}
}
\description{
Defince scc style for ggplot2
}
\details{
Defines a ggplot2 style using SCC font, ensuring consistent
font size styling, colours and plot margin
}
