% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_barchart.r
\name{scc_barchart_grouped}
\alias{scc_barchart_grouped}
\title{Define a grouped barchart in scc colours}
\usage{
scc_barchart_grouped(df_to_plot, x, y, group, title = NULL, subtitle = NULL)
}
\arguments{
\item{df_to_plot}{The dataframe provided}

\item{x}{The x-axis variable}

\item{y}{The y-axis variable}

\item{group}{the variable by which the chart should be grouped}

\item{title}{The title that should go above the plot}

\item{subtitle}{The subtitle that should go below the title}
}
\description{
Define a grouped barchart in scc colours
}
