% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_barchart.r
\name{scc_barchart_pct}
\alias{scc_barchart_pct}
\title{Define barchart with \% values in scc colours}
\usage{
scc_barchart_pct(df_to_plot, x, y, title = NULL, subtitle = NULL)
}
\arguments{
\item{df_to_plot}{The dataframe provided}

\item{x}{The x-axis variable}

\item{y}{The y-axis variable}

\item{title}{The title that should go above the plot}

\item{subtitle}{The subtitle that should go below the title}
}
\description{
Define barchart with \% values in scc colours
}
